Refactoring the solution to use Cilium
I found unusual behavior that seems to point to MetalLB, so I decided to
switch to Cilium; although the problems are still present, so I need to
dig more.

I reduced the scope of the metrics to be pushed to Mimir, but it looks
like that's not the problem.

As a bonus, I enable the Mimir dashboards via Helm, so there is no need
to deploy them separately.